+++ src/ids-base/ids-events-mixin.js
@@ -48,9 +48,14 @@ const IdsEventsMixin = (superclass) => class extends superclass {
 
   /**
    * Detach all event handlers
+   * @param {string} [eventName] an optional event name to filter with
    */
-  detachAllEvents() {
+  detachAllEvents(eventName = undefined) {
+    const doCheck = typeof eventName === 'string' && eventName.length;
     this.handledEvents.forEach((value, key) => {
+      if (doCheck && key !== eventName) {
+        return;
+      }
       this.removeEventListener(key, value.target, value.options);
     });
   }