+++ src/ids-popup/ids-popup.js
@@ -25,6 +25,9 @@ const ALIGNMENTS_Y = [CENTER, 'top', 'bottom'];
 const ALIGNMENTS_EDGES_X = ALIGNMENTS_X.filter((x) => x !== CENTER);
 const ALIGNMENTS_EDGES_Y = ALIGNMENTS_Y.filter((y) => y !== CENTER);
 
+// Arrow Directions (defaults to 'none')
+const ARROW_TYPES = ['none', 'bottom', 'top', 'left', 'right'];
+
 // Types of Popups
 const TYPES = ['none', 'menu', 'menu-alt', 'tooltip', 'tooltip-alt'];
 
@@ -36,6 +39,8 @@ const POPUP_PROPERTIES = [
   'align-y',
   'align-edge',
   'align-target',
+  'arrow',
+  'arrow-target',
   props.ANIMATED,
   props.TYPE,
   props.VISIBLE,
@@ -87,6 +92,10 @@ class IdsPopup extends mix(IdsElement).with(IdsRenderLoopMixin, IdsResizeMixin, 
     this.coords = {
       x: 0,
       y: 0
+    };
+    this.state = {
+      arrow: ARROW_TYPES[0],
+      arrowTarget: null,
     };
     this.isVisible = false;
     this.isAnimated = false;
@@ -399,6 +408,83 @@ class IdsPopup extends mix(IdsElement).with(IdsRenderLoopMixin, IdsResizeMixin, 
   }
 
   /**
+   * Specifies whether to show the Popup Arrow, and in which direction
+   * @param {string} val the arrow direction.  Defaults to `none`
+   */
+  set arrow(val) {
+    let trueVal = ARROW_TYPES[0];
+    if (val && ARROW_TYPES.includes(val)) {
+      trueVal = val;
+    }
+    if (trueVal !== ARROW_TYPES[0]) {
+      this.safeSetAttribute('arrow', `${trueVal}`);
+    } else {
+      this.safeRemoveAttribute('arrow');
+    }
+    this.refresh();
+  }
+
+  /**
+   * @returns {string|null} the arrow setting, or null
+   */
+  get arrow() {
+    const attr = this.getAttribute('arrow');
+    if (!attr) {
+      return ARROW_TYPES[0];
+    }
+    return attr;
+  }
+
+  /**
+   * @readonly
+   * @returns {HTMLElement} referencing the internal arrow element
+   */
+  get arrowEl() {
+    return this.container.querySelector('.arrow');
+  }
+
+  /**
+   * Sets the element to align with via a css selector
+   * @param {any} val ['string|HTMLElement'] a CSS selector string
+   */
+  // @ts-ignore
+  set arrowTarget(val) {
+    const isString = typeof val === 'string' && val.length;
+    const isElem = val instanceof HTMLElement;
+
+    if (!isString && !isElem) {
+      this.state.arrowTarget = undefined;
+      this.removeAttribute('arrow-target');
+      this.refresh();
+      return;
+    }
+
+    let elem;
+    if (isString) {
+      // @TODO Harden for security (XSS)
+      elem = document.querySelector(val);
+      if (!(elem instanceof HTMLElement)) {
+        return;
+      }
+      this.setAttribute('arrow-target', val);
+    } else {
+      elem = val;
+    }
+
+    this.state.arrowTarget = elem;
+    this.refresh();
+  }
+
+  /**
+   * @returns {HTMLElement} the element in the page that the Popup will take
+   * coordinates from for relative placement
+   */
+  // @ts-ignore
+  get arrowTarget() {
+    return this.state.arrowTarget || this.alignTarget;
+  }
+
+  /**
    * The style of popup to use between 'none', 'menu', 'menu-alt', 'tooltip', 'tooltip-alt'
    * @param {string} val The popup type
    */
@@ -502,9 +588,23 @@ class IdsPopup extends mix(IdsElement).with(IdsRenderLoopMixin, IdsResizeMixin, 
 
     // Make the popup actually render before doing placement calcs
     if (this.isVisible) {
-      this.container.classList.add('visible');
+      thisCl.add('visible');
     } else {
-      this.container.classList.remove('open');
+      thisCl.remove('open');
+    }
+
+    // Show/Hide Arrow class, if applicable
+    const arrowClass = this.arrow;
+    const arrowElCl = this.arrowEl.classList;
+    ARROW_TYPES.forEach((type) => {
+      if (type !== 'none' && type !== arrowClass) {
+        arrowElCl.remove(type);
+        this.arrowEl.hidden = true;
+      }
+    });
+    if (this.arrow !== 'none' && !arrowElCl.contains(this.arrow)) {
+      arrowElCl.add(this.arrow);
+      this.arrowEl.hidden = false;
     }
 
     // If no alignment target is present, do a simple x/y coordinate placement.
@@ -547,6 +647,9 @@ class IdsPopup extends mix(IdsElement).with(IdsRenderLoopMixin, IdsResizeMixin, 
       duration: 70,
       timeoutCallback: () => {
         if (this.isVisible) {
+          // If an arrow is displayed, place it correctly.
+          this.placeArrow();
+
           // Always fire the 'show' event
           this.trigger('show', this, {
             bubbles: true,
@@ -717,6 +820,76 @@ class IdsPopup extends mix(IdsElement).with(IdsRenderLoopMixin, IdsResizeMixin, 
   }
 
   /**
+   * Handles alignment of an optional arrow element.  If an arrow target is specified,
+   * the arrow is placed to align correctly against the target.
+   * @returns {void}
+   */
+  placeArrow() {
+    const arrow = this.arrow;
+    const arrowEl = this.arrowEl;
+    const element = this.alignTarget;
+    const target = this.arrowTarget;
+
+    if (arrow === 'none' || !element || !target) {
+      arrowEl.hidden = true;
+      return;
+    }
+
+    // Clear previous styles
+    arrowEl.removeAttribute('hidden');
+    arrowEl.style.marginLeft = '';
+    arrowEl.style.marginTop = '';
+
+    const arrowRect = arrowEl.getBoundingClientRect();
+    const elementRect = element.getBoundingClientRect();
+    const targetRect = target.getBoundingClientRect();
+
+    const newArrowRect = {};
+    const targetMargin = (arrow === 'right' || arrow === 'left') ? 'marginTop' : 'marginLeft';
+
+    let arrowHidden = false;
+    let targetCenter = 0;
+    let currentArrowCenter = 0;
+    let d;
+
+    // Figure out the distance needed to move the arrow to match the position of the `target`
+    if (arrow === 'left' || arrow === 'right') {
+      targetCenter = targetRect.top + (targetRect.height / 2);
+      currentArrowCenter = arrowRect.top + (arrowRect.height / 2);
+      d = targetCenter - currentArrowCenter;
+      newArrowRect.top = arrowRect.top + d;
+      newArrowRect.bottom = arrowRect.bottom + d;
+
+      /* istanbul ignore next */
+      if (newArrowRect.top <= elementRect.top || newArrowRect.bottom >= elementRect.bottom) {
+        arrowHidden = true;
+      }
+    }
+    if (arrow === 'top' || arrow === 'bottom') {
+      targetCenter = targetRect.left + (targetRect.width / 2);
+      currentArrowCenter = arrowRect.left + (arrowRect.width / 2);
+      d = targetCenter - currentArrowCenter;
+      newArrowRect.left = arrowRect.left + d;
+      newArrowRect.right = arrowRect.right + d;
+
+      /* istanbul ignore next */
+      if (newArrowRect.left <= elementRect.left || newArrowRect.right >= elementRect.right) {
+        arrowHidden = true;
+      }
+    }
+
+    // Round the number up
+    d = Math.ceil(d);
+
+    // Hide the arrow if it goes beyond the element boundaries
+    /* istanbul ignore next */
+    if (arrowHidden) {
+      arrowEl.hidden = true;
+    }
+    arrowEl.style[targetMargin] = `${d}px`;
+  }
+
+  /**
    * Turns off the ability of the popup to respond to attribute changes, in order to
    * set an attribute that may incorrectly change the popup's display/state otherwise.
    * @param {string} attr the attribute of the popup that will change, passed to `setAttribute`
@@ -755,6 +928,7 @@ class IdsPopup extends mix(IdsElement).with(IdsRenderLoopMixin, IdsResizeMixin, 
    */
   template() {
     return `<div class="ids-popup">
+      <div class="arrow"></div>
       <div class="content-wrapper">
         <slot name="content"></slot>
       </div>